+++ src/index.ts
@@ -11,14 +11,21 @@ import ICalCalendar, {ICalCalendarData} from './calendar';
  * Create a new, empty calendar and returns it.
  *
  * ```javascript
- * const ical = require('ical-generator');
+ * import ical from 'ical-generator';
+ *
+ * // or use require:
+ * // const ical = require('ical-generator');
+ *
  * const cal = ical();
  * ```
  *
  * You can pass options to setup your calendar or use setters to do this.
  *
  * ```javascript
- * const ical = require('ical-generator');
+ * import ical from 'ical-generator';
+ *
+ * // or use require:
+ * // const ical = require('ical-generator');
  * const cal = ical({domain: 'sebbo.net'});
  *
  * // is the same as
@@ -33,9 +40,11 @@ import ICalCalendar, {ICalCalendarData} from './calendar';
  *
  * @param data Calendar data
  */
-export default function (data?: ICalCalendarData): ICalCalendar {
+function ical(data?: ICalCalendarData): ICalCalendar {
     return new ICalCalendar(data);
 }
+
+export default ical;
 
 export {
     default as ICalAlarm,
@@ -88,3 +97,15 @@ export {
     ICalWeekday,
     ICalTimezone
 } from './types';
+
+export {
+    formatDate,
+    formatDateTZ,
+    escape,
+    foldLines
+} from './tools';
+
+/* istanbul ignore else */
+if (typeof module !== 'undefined') {
+    module.exports = Object.assign(ical, module.exports);
+}